refactor: move models to a new package and restore model definitions
